+++ src/main/java/bookstore/back/io/book/BookUpdateRequest.java
@@ -6,10 +6,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class BookUpdateRequest {
-    private Long id;
+    private Integer id;
     private String name;
     private String author;
     private Integer launchDate;
     private Integer totalQuantity;
-    private Long publisherId;
+    private Integer publisherId;
 }
