+++ src/main/java/bookstore/back/services/RentalService.java
@@ -21,5 +21,5 @@ public interface RentalService {
 
     List<RentalResponseRequest> getPending();
 
-    RentalEntity findById(Long id);
+    RentalEntity findById(Integer id);
 }
