+++ src/main/java/bookstore/back/validation/UserValidation.java
@@ -8,5 +8,5 @@ public interface UserValidation {
 
     void validateUpdate(UserEntity user);
 
-    void validateForDelete(Long id);
+    void validateForDelete(Integer id);
 }
